feat(react-components): refactoring rule based outputs functions and main component
